initail project

+++ src/main/java/com/arjuncodes/studentsystem/repository/StudentRepository.java
@@ -1,9 +1,9 @@
 package com.arjuncodes.studentsystem.repository;
 
-import com.arjuncodes.studentsystem.model.Student;
+import com.arjuncodes.studentsystem.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student,Integer> {
+public interface StudentRepository extends JpaRepository<Student,Long> {
 }
